# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/api/academia/professors": {
            "get": {
                "summary": "Get Professor",
                "operationId": "get_professor_api_academia_professors_get",
                "parameters": [
                    {
                        "name": "grad_didactic",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Grad Didactic"
                        }
                    },
                    {
                        "name": "nume",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Nume"
                        }
                    },
                    {
                        "name": "prenume",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Prenume"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Email"
                        }
                    },
                    {
                        "name": "tip_asociere",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Tip Asociere"
                        }
                    },
                    {
                        "name": "afiliere",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Afiliere"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "default": 20,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProfesorPydantic"
                                    },
                                    "title": "Response Get Professor Api Academia Professors Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/academia/students": {
            "get": {
                "summary": "Get Student",
                "operationId": "get_student_api_academia_students_get",
                "parameters": [
                    {
                        "name": "nume",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Nume"
                        }
                    },
                    {
                        "name": "prenume",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Prenume"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Email"
                        }
                    },
                    {
                        "name": "ciclu_studii",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Ciclu Studii"
                        }
                    },
                    {
                        "name": "an_studiu",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "An Studiu"
                        }
                    },
                    {
                        "name": "grupa",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Grupa"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "default": 20,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StudentPydantic"
                                    },
                                    "title": "Response Get Student Api Academia Students Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/academia/lectures": {
            "get": {
                "summary": "Get Lecture",
                "operationId": "get_lecture_api_academia_lectures_get",
                "parameters": [
                    {
                        "name": "cod",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Cod"
                        }
                    },
                    {
                        "name": "id_titular",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Id Titular"
                        }
                    },
                    {
                        "name": "nume_disciplina",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Nume Disciplina"
                        }
                    },
                    {
                        "name": "an_studiu",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "An Studiu"
                        }
                    },
                    {
                        "name": "tip_disciplina",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Tip Disciplina"
                        }
                    },
                    {
                        "name": "categorie_disciplina",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Categorie Disciplina"
                        }
                    },
                    {
                        "name": "tip_examinare",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Tip Examinare"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "default": 20,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DisciplinaPydantic"
                                    },
                                    "title": "Response Get Lecture Api Academia Lectures Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/academia/professors/{id}/lectures": {
            "get": {
                "summary": "Get Professor Lectures",
                "operationId": "get_professor_lectures_api_academia_professors__id__lectures_get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/academia/students/{id}/lectures": {
            "get": {
                "summary": "Get Student Lectures",
                "operationId": "get_student_lectures_api_academia_students__id__lectures_get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "DisciplinaPydantic": {
                "properties": {
                    "cod": {
                        "type": "string",
                        "title": "Cod"
                    },
                    "id_titular": {
                        "type": "integer",
                        "title": "Id Titular"
                    },
                    "nume_disciplina": {
                        "type": "string",
                        "title": "Nume Disciplina"
                    },
                    "an_studiu": {
                        "type": "integer",
                        "title": "An Studiu"
                    },
                    "tip_disciplina": {
                        "type": "string",
                        "title": "Tip Disciplina"
                    },
                    "categorie_disciplina": {
                        "type": "string",
                        "title": "Categorie Disciplina"
                    },
                    "tip_examinare": {
                        "type": "string",
                        "title": "Tip Examinare"
                    }
                },
                "type": "object",
                "required": [
                    "cod",
                    "id_titular",
                    "nume_disciplina",
                    "an_studiu",
                    "tip_disciplina",
                    "categorie_disciplina",
                    "tip_examinare"
                ],
                "title": "DisciplinaPydantic"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ProfesorPydantic": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "nume": {
                        "type": "string",
                        "title": "Nume"
                    },
                    "prenume": {
                        "type": "string",
                        "title": "Prenume"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "grad_didactic": {
                        "type": "string",
                        "title": "Grad Didactic"
                    },
                    "tip_asociere": {
                        "type": "string",
                        "title": "Tip Asociere"
                    },
                    "afiliere": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Afiliere"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "nume",
                    "prenume",
                    "email",
                    "grad_didactic",
                    "tip_asociere",
                    "afiliere"
                ],
                "title": "ProfesorPydantic"
            },
            "StudentPydantic": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "nume": {
                        "type": "string",
                        "title": "Nume"
                    },
                    "prenume": {
                        "type": "string",
                        "title": "Prenume"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "ciclu_studii": {
                        "type": "string",
                        "title": "Ciclu Studii"
                    },
                    "an_studiu": {
                        "type": "integer",
                        "title": "An Studiu"
                    },
                    "grupa": {
                        "type": "string",
                        "title": "Grupa"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "nume",
                    "prenume",
                    "email",
                    "ciclu_studii",
                    "an_studiu",
                    "grupa"
                ],
                "title": "StudentPydantic"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}